+++ src/main/java/com/duoer/campus/dao/FeedingRecordTempMapper.java
@@ -13,5 +13,5 @@ public interface FeedingRecordTempMapper extends BaseMapper<FeedingRecordTemp> {
      */
     @Update("update feeding_record_tmp set deleted = 1, version = version + 1 " +
             "where record_id = #{id} and version = #{version} and deleted = 0")
-    int myDeleteById(@Param("id") int id, @Param("version") int version);
+    int myDeleteById(@Param("id") long id, @Param("version") int version);
 }
