+++ src/main/java/com/duoer/campus/dto/AppearanceRecordDTO.java
@@ -7,11 +7,11 @@ import lombok.*;
 @Data
 @NoArgsConstructor
 public class AppearanceRecordDTO extends RecordDTO {
-    private Integer formerId;
+    private Long formerId;
 
-    public AppearanceRecordDTO(Integer recordId, Integer catId, String catName,
-                               String username, Integer locationId, String location,
-                               String recordTime, String lastUpdate, Integer formerId) {
+    public AppearanceRecordDTO(Long recordId, Long catId, String catName,
+                               String username, Long locationId, String location,
+                               String recordTime, String lastUpdate, Long formerId) {
         super(recordId, catId, catName, username, locationId, location, recordTime, lastUpdate);
         this.formerId = formerId;
     }
