+++ src/main/java/com/duoer/campus/entity/AppearanceRecord.java
@@ -10,8 +10,8 @@ import lombok.ToString;
 @NoArgsConstructor
 @Data
 public class AppearanceRecord extends MyRecord {
-    public AppearanceRecord(Integer recordId, Integer catId, String username,
-                            Integer locationId, String recordTime, String lastUpdate) {
+    public AppearanceRecord(Long recordId, Long catId, String username,
+                            Long locationId, String recordTime, String lastUpdate) {
         super(recordId, catId, username, locationId, recordTime, lastUpdate);
     }
 }
