+++ src/main/java/com/duoer/campus/entity/AppearanceRecordTemp.java
@@ -14,13 +14,13 @@ import lombok.ToString;
 @Data
 @TableName("appearance_record_tmp")
 public class AppearanceRecordTemp extends AppearanceRecord {
-    private Integer formerId;
+    private Long formerId;
     @TableLogic(value = "0", delval = "1")
     private Integer deleted;
     @Version
     private Integer version;
 
-    public AppearanceRecordTemp(AppearanceRecord ar, Integer formerId) {
+    public AppearanceRecordTemp(AppearanceRecord ar, Long formerId) {
         super(null, ar.getCatId(), ar.getUsername(), ar.getLocationId(),
                 ar.getRecordTime(), ar.getLastUpdate());
         this.formerId = formerId;
