+++ src/main/java/com/duoer/campus/service/RecordService.java
@@ -45,7 +45,7 @@ public interface RecordService {
      * @param type 类型
      * @return 记录
      */
-    RecordDTO getRecordById(int id, String type);
+    RecordDTO getRecordById(long id, String type);
 
     /**
      * 按编号查临时表记录
@@ -54,7 +54,7 @@ public interface RecordService {
      * @param type 类型
      * @return 记录
      */
-    RecordDTO getTempRecordById(int id, String type);
+    RecordDTO getTempRecordById(long id, String type);
 
     /**
      * 获取某猫咪所有记录
@@ -63,7 +63,7 @@ public interface RecordService {
      * @param type 类型
      * @return 所有记录集合
      */
-    List<? extends RecordDTO> getCatOwnRecords(int id, String type);
+    List<? extends RecordDTO> getCatOwnRecords(long id, String type);
 
     /**
      * 添加新记录
@@ -88,7 +88,7 @@ public interface RecordService {
      * @param type 类型
      * @return 状态码
      */
-    int addRecordCheckPass(int id, String type);
+    int addRecordCheckPass(long id, String type);
 
     /**
      * 修改某记录
@@ -107,7 +107,7 @@ public interface RecordService {
      * @param isAdmin 是否管理员
      * @return 状态码
      */
-    int deleteRecord(int[] ids, String type, String username, Boolean isAdmin);
+    int deleteRecord(long[] ids, String type, String username, Boolean isAdmin);
 
     /**
      * 删除某些临时表记录
@@ -116,5 +116,5 @@ public interface RecordService {
      * @param type 类型
      * @return 状态码
      */
-    int deleteTempRecord(List<Integer> ids, String type);
+    int deleteTempRecord(List<Long> ids, String type);
 }
